support splitchapters command for youtube

+++ youtube_utils.py
@@ -16,23 +16,35 @@ from settings import get_default_logger
 from utils import cache_path_for_mp3_url, run_command
 
 
-def set_metadata_from_info_file(mp3_path: Path) -> None:
-    info_json_path = [
+def find_info_json_file_path(mp3_path: Path) -> Path:
+    return [
         p
         for p in mp3_path.parent.glob(f"{mp3_path.stem}.*")
         if p.name.endswith(".info.json")
     ][0]
 
+
+def get_chapter_names_from_info_json_file(info_json_path: Path) -> list[str]:
+    found_metadata = json.loads(info_json_path.read_text())
+    chapters = found_metadata["chapters"]
+    chapters.sort(key=lambda c: c["start_time"])
+    return [c["title"] for c in chapters]
+
+
+def set_metadata_from_info_file(
+    mp3_path: Path, info_json_path: Path, title: str | None = None
+) -> None:
     found_metadata = json.loads(info_json_path.read_text())
 
-    mp3_utils.change_metadata(mp3_path, "title", found_metadata["title"])
+    if title is None:
+        title = found_metadata["title"]
+
+    mp3_utils.change_metadata(mp3_path, "title", title)
     if "album" in found_metadata:
         mp3_utils.change_metadata(mp3_path, "album", found_metadata["album"])
 
-    info_json_path.unlink()
 
-
-def merge_mp3_with_cover(mp3_path: Path) -> None:
+def get_thumbnail_path_for_mp3(mp3_path: Path) -> Path | None:
     thumbnails = [
         p
         for p in mp3_path.parent.glob(f"{mp3_path.stem}.*")
@@ -41,7 +53,7 @@ def merge_mp3_with_cover(mp3_path: Path) -> None:
 
     if len(thumbnails) == 0:
         get_default_logger().info(f"No thumbnails found for {mp3_path}")
-        return
+        return None
 
     if any(
         desired_format_thumbnails := [
@@ -56,18 +68,28 @@ def merge_mp3_with_cover(mp3_path: Path) -> None:
     assert thumbnail.suffix[1:] == DESIRED_THUMBNAIL_FORMAT
 
     crop_image_to_square(thumbnail)
-    mp3_utils.set_cover(mp3_path, thumbnail)
 
     for thumbnail_filepath in thumbnails:
-        os.remove(thumbnail_filepath)
+        if thumbnail_filepath != thumbnail:
+            os.remove(thumbnail_filepath)
+
+    return thumbnail
 
 
-def ytdl_download_song(url: str) -> Path:
+def ytdl_download_song(url: str, split_chapters: bool) -> list[Path]:
     output_filepath = cache_path_for_mp3_url(url)
 
     get_default_logger().info(
         f"Downloading youtube audio from url: {url} with filename {output_filepath}"
     )
+
+    output_filepath_str = output_filepath.as_posix()
+
+    split_chapters_args = [
+        "--split-chapters",
+        "--output",
+        f"chapter:{output_filepath_str}.%(section_number)s.chapter.mp3",
+    ]
 
     run_command(
         [
@@ -77,12 +99,14 @@ def ytdl_download_song(url: str) -> Path:
             "--extract-audio",
             "--write-thumbnail",
             "--write-info-json",
+            "--no-write-comments",
             "--audio-format",
             "mp3",
             "--audio-quality",
             "0",
             "--output",
-            output_filepath.as_posix(),
+            output_filepath_str,
+            *(split_chapters_args if split_chapters else []),
             url,
         ],
     )
@@ -90,10 +114,29 @@ def ytdl_download_song(url: str) -> Path:
     assert output_filepath.exists()
     get_default_logger().info(f"Audio from url {url} downloaded")
 
-    merge_mp3_with_cover(output_filepath)
-    set_metadata_from_info_file(output_filepath)
+    info_json_file = find_info_json_file_path(output_filepath)
+    thumbnail_image_file = get_thumbnail_path_for_mp3(output_filepath)
 
-    return output_filepath
+    if not split_chapters:
+        if thumbnail_image_file is not None:
+            mp3_utils.set_cover(output_filepath, thumbnail_image_file)
+        set_metadata_from_info_file(output_filepath, info_json_file)
+        results = [output_filepath]
+    else:
+        results = []
+        chapter_titles = get_chapter_names_from_info_json_file(info_json_file)
+        for chapter_idx, chapter_title in enumerate(chapter_titles, start=1):
+            filename = Path(output_filepath_str + f".{chapter_idx}.chapter.mp3")
+            if thumbnail_image_file is not None:
+                mp3_utils.set_cover(filename, thumbnail_image_file)
+            set_metadata_from_info_file(filename, info_json_file, title=chapter_title)
+            results.append(filename)
+
+    info_json_file.unlink()
+    if thumbnail_image_file is not None:
+        thumbnail_image_file.unlink()
+
+    return results
 
 
 def extract_youtube_id(link: str) -> str:
@@ -114,7 +157,8 @@ def extract_youtube_id(link: str) -> str:
 
 def playlist_url_to_video_urls(playlist_url: str) -> list[str]:
     p1 = run_command(
-        ["yt-dlp", "--skip-download", "--flat-playlist", playlist_url, "-j"], allow_errors=True
+        ["yt-dlp", "--skip-download", "--flat-playlist", playlist_url, "-j"],
+        allow_errors=True,
     )
     p2 = run_command(["jq", "-r", ".webpage_url"], stdin=p1.stdout)
